+++ library/src/main/kotlin/eu/tib/orkg/prototype/statements/domain/model/PaperResourceWithPath.kt
@@ -1,0 +1,8 @@
+package eu.tib.orkg.prototype.statements.domain.model
+
+typealias Path = List<List<Thing>>
+
+data class PaperResourceWithPath(
+    val paper: Resource,
+    var path: Path,
+)